monitoring-observation/app/pyproject.toml
[project]
name = "litter-detection-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "eclipse-zenoh>=1.7.2",
    "pydantic>=2.0",
    "opentelemetry-distro>=0.45",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45",
    "opentelemetry-instrumentation-sqlalchemy>=0.45",
    "opentelemetry-instrumentation-logging>=0.45",
]
